refactor: refactored several coponents and merged files  within the searching-consultant dir.
