+++ components/searching-consultant/filters/filterCategories/PriceRangeModal.js
@@ -3,7 +3,7 @@ import { Text, StyleSheet, View } from 'react-native';
 import Slider from '@react-native-community/slider';
 
 import { useConsultantFiltersContext } from '../../../../context/ConsultantFilterContext';
-import ModalContent from './FilterModalContainer';
+import FilterCommonContent from '../FilterCommonContent';
 import tabs from '../../../../locales/en/Tabs.json';
 
 const PriceRangeModal = ({ closeModal }) => {
@@ -32,7 +32,7 @@ const PriceRangeModal = ({ closeModal }) => {
 	);
 
 	return (
-		<ModalContent
+		<FilterCommonContent
 			closeModal={closeModal}
 			headerAction={() => setPrice(0)}
 			headerFilter={tabs.PriceRange.label}
